third_party/blink/web_tests/platform/generic/fast/forms/input-common/duplicate-names-expected.txt

Multiple radio buttons/checkboxes with the same name should map to a countable collection object when accessed through either document.form.name or document.getElementsbyName(). There is no standardization for what that object is called (Collection, NodeList, HTMLCollection, etc.), nor is it always the same for both methods of access.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Radio button: Male Female
Checkbox: 1
Checkbox: 1 2


PASS typeof document.f.cb0 is 'undefined'
PASS document.f.cb1.nodeName is 'INPUT'
PASS document.f.cb2.length is 2
PASS document.f.rb2.length is 2
PASS document.getElementsByName('cb2').length is 2
PASS document.getElementsByName('rb2').length is 2
PASS successfullyParsed is true

TEST COMPLETE
